pyproject.toml
[tool.poetry]
name = "smartflow"
version = "0.1.0"
description = "A modular AI workflow engine with prompt feedback loops, LLM/non-LLM tasks, and token tracking"
authors = ["Sanket Naik <sanketn26@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/sanketn26/smartflow"
repository = "https://github.com/sanketn26/smartflow"
keywords = ["ai", "workflow", "llm", "automation", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [{include = "smartflow"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain-core = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
langchain = "^0.3.0"
faiss-cpu = "^1.7.4"
streamlit = "^1.28.0"
python-dotenv = "^1.0.0"
boto3 = "^1.34.0"
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
isort = "^5.12.0"
pre-commit = "^3.4.0"
safety = "^2.3.0"
bandit = "^1.7.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.0"
mkdocs-material = "^9.2.0"
mkdocstrings = "^0.22.0"

[tool.poetry.scripts]
smartflow = "smartflow.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
